+++ lib/jdbc-helper/connector/sql_server_connector.rb
@@ -2,8 +2,8 @@
 # Junegunn Choi (junegunn.c@gmail.com)
 
 module JDBCHelper
-# Shortcut connector for PostgreSQL
-class PostgresConnector < Connector
+# Shortcut connector for MS SQL Server
+class SqlServerConnector < Connector
   include Constants
   include Constants::Connector
 
@@ -16,7 +16,7 @@ class PostgresConnector < Connector
   # @return [JDBCHelper::Connection]
   def self.connect(host, user, password, db,
            timeout = DEFAULT_LOGIN_TIMEOUT, 
-           extra_params = DEFAULT_PARAMETERS[:postgres], &block)
+           extra_params = DEFAULT_PARAMETERS[:sqlserver], &block)
 
     if extra_params && extra_params.is_a?(Hash) == false
       raise ArgumentError.new('extra_params must be a hash')
@@ -24,14 +24,14 @@ class PostgresConnector < Connector
 
     conn = Connection.new(
       (extra_params || {}).merge(
-        :driver   => JDBC_DRIVER[:postgres],
-        :url      => "jdbc:postgresql://#{host}/#{db}",
+        :driver   => JDBC_DRIVER[:sqlserver],
+        :url      => "jdbc:sqlserver://#{host};databaseName=#{db};",
         :user     => user,
         :password => password,
         :timeout  => timeout))
 
     block_given? ? ensure_close(conn, &block) : conn
   end
-end#PostgresConnector
+end#SqlServerConnector
 end#JDBCHelper