+++ lib/jdbc-helper/sql/expression.rb
@@ -8,8 +8,9 @@ class SQL
   # Generate SQL snippet, prevents the string from being quoted.
   # @param [String] SQL snippet
   # @return [JDBCHelper::SQL::Expression]
+  # @since 0.7.0
   def self.expr sql
-    SimpleExpression.new sql
+    ScalarExpression.new sql
   end
 
   # "is not null" expression for where clauses
@@ -23,36 +24,42 @@ class SQL
 
   # Greater-than expression for where clauses
   # @return [JDBCHelper::SQL::Expression]
+  # @since 0.7.0
   def self.gt v
     ParameterizedExpression.new '>', v
   end
 
   # Less-than expression for where clauses
   # @return [JDBCHelper::SQL::Expression]
+  # @since 0.7.0
   def self.lt v
     ParameterizedExpression.new '<', v
   end
 
   # Less-than-or-equal-to expression for where clauses
   # @return [JDBCHelper::SQL::Expression]
+  # @since 0.7.0
   def self.le v
     ParameterizedExpression.new '<=', v
   end
 
   # Greater-than-or-equal-to expression for where clauses
   # @return [JDBCHelper::SQL::Expression]
+  # @since 0.7.0
   def self.ge v
     ParameterizedExpression.new '>=', v
   end
 
   # Not-equal expression for where clauses
   # @return [JDBCHelper::SQL::Expression]
+  # @since 0.7.0
   def self.ne v
     ParameterizedExpression.new '<>', v
   end
 
   # Like expression for where clauses
   # @return [JDBCHelper::SQL::Expression]
+  # @since 0.7.0
   def self.like v
     raise ArgumentError.new('Like expression must be given as a String') unless v.is_a?(String)
     ParameterizedExpression.new 'like', v
@@ -60,11 +67,13 @@ class SQL
 
   # "Not like" expression for where clauses
   # @return [JDBCHelper::SQL::Expression]
+  # @since 0.7.0
   def self.not_like v
     raise ArgumentError.new('Like expression must be given as a String') unless v.is_a?(String)
     ParameterizedExpression.new 'not like', v
   end
 
+  # @since 0.7.0
   class Expression
     def initialize
       raise Exception.new("JDBCHelper::SQL::Expression is an abstract class")
@@ -83,7 +92,8 @@ class SQL
     end
   end
 
-  class SimpleExpression < Expression
+  # @since 0.7.0
+  class ScalarExpression < Expression
     def initialize sql
       @sql = SQL.check sql.to_s
     end
@@ -97,6 +107,7 @@ class SQL
     end
   end
 
+  # @since 0.7.0
   class NotNullExpression < Expression
     def self.singleton
       @@singleton ||= NotNullExpression.new
@@ -114,6 +125,7 @@ class SQL
     end
   end
 
+  # @since 0.7.0
   class ParameterizedExpression < Expression
     def initialize operator, param
       @operator = operator