+++ lib/jdbc-helper/wrapper/sequence_wrapper.rb
@@ -15,16 +15,29 @@ class SequenceWrapper < ObjectWrapper
   # @return [String]
   alias to_s name
 
+  def initialize(conn, name)
+    super conn, name
+
+    case conn.driver
+    when /oracle/
+      @nextval_sql = "select #{name}.nextval from dual"
+      @currval_sql = "select #{name}.currval from dual"
+    else
+      @nextval_sql = "select nextval('#{name}')"
+      @currval_sql = "select currval('#{name}')"
+    end
+  end
+
   # Increments the sequence and returns the value
   # @return [Fixnum]
   def nextval
-    @connection.query("select #{name}.nextval from dual")[0][0].to_i
+    @connection.query(@nextval_sql)[0][0].to_i
   end
 
   # Returns the incremented value of the sequence
   # @return [Fixnum]
   def currval
-    @connection.query("select #{name}.currval from dual")[0][0].to_i
+    @connection.query(@currval_sql)[0][0].to_i
   end
 
   # Recreates the sequence. Cannot be undone.