+++ test/test_connection.rb
@@ -3,14 +3,6 @@ require 'helper'
 class TestConnection < Test::Unit::TestCase
   include JDBCHelperTestHelper
 
-  def setup
-  end
-
-  def teardown
-    conn.update "drop table #{TEST_TABLE}" rescue nil
-    conn.update "drop procedure #{TEST_PROCEDURE}" rescue nil
-  end
-
   TEST_TABLE = 'tmp_jdbc_helper_test'
   TEST_PROCEDURE = 'tmp_jdbc_helper_test_proc'
 
@@ -65,7 +57,7 @@ class TestConnection < Test::Unit::TestCase
     conn.update "drop table #{TEST_TABLE}" rescue nil
     cnt = conn.update "
       create table #{TEST_TABLE} (
-        a timestamp,
+        a #{@type == :sqlserver ? 'datetime' : 'timestamp'},
         b date
         #{", c time" if @type == :mysql}
       )"
@@ -122,11 +114,15 @@ class TestConnection < Test::Unit::TestCase
 
         # initialize with execution block
         conn = JDBCHelper::Connection.new(conn_info) do | c |
+          init_dual! c
+
           c2 = c.clone
           assert c2.java_obj != c.java_obj
 
           c.query('select 1 from dual')
           c2.query('select 1 from dual')
+          drop_dual! c
+
           assert_equal c.closed?, false
 
           c2.close
@@ -328,7 +324,8 @@ class TestConnection < Test::Unit::TestCase
   def test_prepared_update_batch
     each_connection do | conn |
       reset_test_table conn
-      ins = conn.prepare "insert into #{TEST_TABLE} values (?, ?)"
+      # Note: (a, b) required for SQL Server
+      ins = conn.prepare "insert into #{TEST_TABLE} (a, b) values (?, ?)"
       assert_equal 2, ins.parameter_count
       assert_equal conn.prepared_statements.first, ins
 
@@ -344,7 +341,7 @@ class TestConnection < Test::Unit::TestCase
       # add_batch execute_batch
       2.times do |iter|
         reset_test_table conn
-        ins = conn.prepare "insert into #{TEST_TABLE} values (?, ?)"
+        ins = conn.prepare "insert into #{TEST_TABLE} (a, b) values (?, ?)"
         assert_equal conn.prepared_statements.first, ins if iter == 0
 
         count.times do | p |
@@ -363,7 +360,7 @@ class TestConnection < Test::Unit::TestCase
 
       # add_batch clear_batch
       reset_test_table conn
-      ins = conn.prepare "insert into #{TEST_TABLE} values (?, ?)"
+      ins = conn.prepare "insert into #{TEST_TABLE} (a, b) values (?, ?)"
       assert_equal conn.prepared_statements.last, ins # count is first
 
       # clear_batch
@@ -466,7 +463,11 @@ class TestConnection < Test::Unit::TestCase
       ts = Time.now
       conn.prepare("insert into #{TEST_TABLE} (a) values (?)").update(ts)
       got = conn.query("select a from #{TEST_TABLE}")[0][0]
-      assert [ts.to_i * 1000, (ts.to_f * 1000).to_i].include?(got.getTime)
+      assert [
+              ts.to_i * 1000, 
+              (ts.to_f * 1000).to_i,
+              (ts.to_f * 1000).to_i * 10 / 10   # SQL Server seems to round up the millisecond precision
+             ].include?(got.getTime)
     end
   end
 
@@ -474,6 +475,8 @@ class TestConnection < Test::Unit::TestCase
   # Oracle and MySQL behave differently.
   def test_callable_statement
     each_connection do | conn |
+      next unless [:mysql, :oracle].include?(@type) # TODO
+
       # Creating test procedure (Defined in JDBCHelperTestHelper)
       create_test_procedure conn, TEST_PROCEDURE